updated uptime-server
